created card component and renamed other components
